+++ src/index.js
@@ -1,7 +1,6 @@
 import apiCreate from './api/create';
 import apiEmit from './api/emit';
 import apiInit from './api/init';
-import apiNoConflict from './api/no-conflict';
 import apiVersion from './api/version';
 import assign from './util/assign';
 import assignSafe from './util/assign-safe';
@@ -104,11 +103,7 @@ function skate (id, userOptions) {
 skate.create = apiCreate;
 skate.emit = apiEmit;
 skate.init = apiInit;
-skate.noConflict = apiNoConflict;
 skate.version = apiVersion;
-
-// Global
-window.skate = skate;
 
 // ES6
 export default skate;